minor refactorings addressing review comments
